PrivacyGuard-AI-CRM-main/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
ollama>=0.3.0
pytz>=2024.1
reportlab>=4.2.0
requests>=2.32.0
